requirements.txt
certifi==2020.12.5
chardet==3.0.4
Click==7.0
dnspython==2.1.0
Flask==1.1.1
Flask-Cors==3.0.10
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==8.1.2
pymongo==3.9.0
requests==2.22.0
six==1.15.0
urllib3==1.25.11
Werkzeug==0.15.6
wincertstore==0.2
